+++ lib/main.dart
@@ -6,6 +6,7 @@ import 'package:turac/home/main_product.dart';
 import 'package:get/get.dart';
 import 'package:turac/home/signup_page.dart';
 import 'package:turac/pages/product/popular_product_detail.dart';
+import 'package:turac/pages/user_profile.dart';
 import 'helper/dependencies.dart' as dep;
 
 // this is the init method for the api's
@@ -32,10 +33,11 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       // ignore: prefer_const_constructors
-      home: MainProductPage(),
+      // home: MainProductPage(),
       // home: PopularProductDetail(),
       // home:LoginPage(),
       // home:SignupPage(),
+      home:AccountProfile(),
     );
   }
 }
