+++ lib/repository/auth_status.dart
@@ -1,133 +1,133 @@
-// ignore_for_file: constant_identifier_names, import_of_legacy_library_into_null_safe, prefer_typing_uninitialized_variables
+// // ignore_for_file: constant_identifier_names, import_of_legacy_library_into_null_safe, prefer_typing_uninitialized_variables
 
-import 'dart:async';
-import 'dart:convert';
-import 'dart:io';
+// import 'dart:async';
+// import 'dart:convert';
+// import 'dart:io';
 
-import 'package:flutter/material.dart';
-import 'package:http/http.dart';
-import 'package:turac/models/users_model.dart';
-import 'package:turac/utils/app_constants.dart';
-
-
-enum Status {
-  NotLoggedIn,
-  NotRegistered,
-  LoggedIn,
-  Registered,
-  Authenticating,
-  Registering,
-  LoggedOut
-}
-
-class AuthProvider with ChangeNotifier {
-
-  Status _loggedInStatus = Status.NotLoggedIn;
-  Status _registeredInStatus = Status.NotRegistered;
-
-  Status get loggedInStatus => _loggedInStatus;
-  Status get registeredInStatus => _registeredInStatus;
-
-  get mobile => null;
+// import 'package:flutter/material.dart';
+// import 'package:http/http.dart';
+// import 'package:turac/models/users_model.dart';
+// import 'package:turac/utils/app_constants.dart';
 
 
-  Future<Map<String, dynamic>> login(String email, String password) async {
-    // ignore: prefer_typing_uninitialized_variables
-    var result;
+// enum Status {
+//   NotLoggedIn,
+//   NotRegistered,
+//   LoggedIn,
+//   Registered,
+//   Authenticating,
+//   Registering,
+//   LoggedOut
+// }
 
-    final Map<String, dynamic> loginData = {
-      'user': {
-        'email': email,
-        'password': password,
-        // 'mobile': mobile
-      }
-    };
+// class AuthProvider with ChangeNotifier {
 
-    _loggedInStatus = Status.Authenticating;
-    notifyListeners();
+//   Status _loggedInStatus = Status.NotLoggedIn;
+//   Status _registeredInStatus = Status.NotRegistered;
 
-    Response response = await post(
-      AppConstants.login_URI,
-      body: json.encode(loginData),
-      headers: {'Content-Type': 'application/json'},
-    );
+//   Status get loggedInStatus => _loggedInStatus;
+//   Status get registeredInStatus => _registeredInStatus;
 
-    if (response.statusCode == 200) {
-      final Map<String, dynamic> responseData = json.decode(response.body);
-
-      var userData = responseData['data'];
-
-      User authUser = User.fromJson(userData);
-
-      UserPreferences().saveUser(authUser);
-
-      _loggedInStatus = Status.LoggedIn;
-      notifyListeners();
-
-      result = {'status': true, 'message': 'Successful', 'user': authUser};
-    } else {
-      _loggedInStatus = Status.NotLoggedIn;
-      notifyListeners();
-      result = {
-        'status': false,
-        'message': json.decode(response.body)['error']
-      };
-    }
-    return result;
-  }
-
-  Future<FutureOr> register(String email, String password, String passwordConfirmation) async {
-
-    final Map<String, dynamic> registrationData = {
-      'user': {
-        'email': email,
-        'password': password,
-        // 'password_confirmation': passwordConfirmation
-      }
-    };
+//   get mobile => null;
 
 
-    _registeredInStatus = Status.Registering;
-    notifyListeners();
+//   Future<Map<String, dynamic>> login(String email, String password) async {
+//     // ignore: prefer_typing_uninitialized_variables
+//     var result;
 
-    return await post(AppConstants.register_URI,
-        body: json.encode(registrationData),
-        headers: {'Content-Type': 'application/json'})
-        .then(onValue)
-        .catchError(onError);
-  }
+//     final Map<String, dynamic> loginData = {
+//       'user': {
+//         'email': email,
+//         'password': password,
+//         // 'mobile': mobile
+//       }
+//     };
 
-  static Future<FutureOr> onValue(Response response) async {
-    var result;
-    final Map<String, dynamic> responseData = json.decode(response.body);
+//     _loggedInStatus = Status.Authenticating;
+//     notifyListeners();
 
-    if (response.statusCode == 200) {
+//     Response response = await post(
+//       AppConstants.login_URI,
+//       body: json.encode(loginData),
+//       headers: {'Content-Type': 'application/json'},
+//     );
 
-      var userData = responseData['data'];
+//     if (response.statusCode == 200) {
+//       final Map<String, dynamic> responseData = json.decode(response.body);
 
-      User authUser = User.fromJson(userData);
+//       var userData = responseData['data'];
 
-      UserPreferences().saveUser(authUser);
-      result = {
-        'status': true,
-        'message': 'Successfully registered',
-        'data': authUser
-      };
-    } else {
+//       User authUser = User.fromJson(userData);
 
-      result = {
-        'status': false,
-        'message': 'Registration failed',
-        'data': responseData
-      };
-    }
+//       UserPreferences().saveUser(authUser);
 
-    return result;
-  }
+//       _loggedInStatus = Status.LoggedIn;
+//       notifyListeners();
 
-  static onError(error) {
-    print("the error is $error.detail");
-    return {'status': false, 'message': 'Unsuccessful Request', 'data': error};
-  }
+//       result = {'status': true, 'message': 'Successful', 'user': authUser};
+//     } else {
+//       _loggedInStatus = Status.NotLoggedIn;
+//       notifyListeners();
+//       result = {
+//         'status': false,
+//         'message': json.decode(response.body)['error']
+//       };
+//     }
+//     return result;
+//   }
 
-}
+//   Future<FutureOr> register(String email, String password, String passwordConfirmation) async {
+
+//     final Map<String, dynamic> registrationData = {
+//       'user': {
+//         'email': email,
+//         'password': password,
+//         // 'password_confirmation': passwordConfirmation
+//       }
+//     };
+
+
+//     _registeredInStatus = Status.Registering;
+//     notifyListeners();
+
+//     return await post(AppConstants.register_URI,
+//         body: json.encode(registrationData),
+//         headers: {'Content-Type': 'application/json'})
+//         .then(onValue)
+//         .catchError(onError);
+//   }
+
+//   static Future<FutureOr> onValue(Response response) async {
+//     var result;
+//     final Map<String, dynamic> responseData = json.decode(response.body);
+
+//     if (response.statusCode == 200) {
+
+//       var userData = responseData['data'];
+
+//       User authUser = User.fromJson(userData);
+
+//       UserPreferences().saveUser(authUser);
+//       result = {
+//         'status': true,
+//         'message': 'Successfully registered',
+//         'data': authUser
+//       };
+//     } else {
+
+//       result = {
+//         'status': false,
+//         'message': 'Registration failed',
+//         'data': responseData
+//       };
+//     }
+
+//     return result;
+//   }
+
+//   static onError(error) {
+//     print("the error is $error.detail");
+//     return {'status': false, 'message': 'Unsuccessful Request', 'data': error};
+//   }
+
+// }
