+++ lib/repository/change_notifier.dart
@@ -1,13 +1,13 @@
-import 'package:flutter/foundation.dart';
-import 'package:turac/models/users_model.dart';
+// import 'package:flutter/foundation.dart';
+// import 'package:turac/models/users_model.dart';
 
-class UserProvider with ChangeNotifier {
-  User _user = new User();
+// class UserProvider with ChangeNotifier {
+//   User _user = new User();
 
-  User get user => _user;
+//   User get user => _user;
 
-  void setUser(User user) {
-    _user = user;
-    notifyListeners();
-  }
-}
+//   void setUser(User user) {
+//     _user = user;
+//     notifyListeners();
+//   }
+// }
