+++ lib/widgets/account_widget.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:turac/utils/dimensions.dart';
+import 'package:turac/widgets/appIcon.dart';
+import 'package:turac/widgets/big_text.dart';
+
+class AccountWidget extends StatelessWidget {
+  AppIcon appIcon;
+  BigText bigText;
+  
+  AccountWidget({Key? key, required this.appIcon, required this.bigText}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.only(left: Dimensions.width20,
+      top: Dimensions.width10,
+      bottom: Dimensions.width10),
+      child: Row(children: [
+        appIcon,
+        bigText
+      ]),
+    );
+  }
+}
